Throw an exception when we get an unexpected End of Stream marker from HttpClient.
This fixes potential data loss when reading from GCS and the connection gets prematurely terminated. Unfortunately it's impossible to make it work for compressed streams at this level, since we can't get a proper count of transferred bytes.

Root cause of the problem: JDK ignores incomplete HTTP responses when Content-Length is specified. Chunked encoded streams work fine and unexpected termination is being detected.

----Internal Notes----

This is a bug in JDK: []

Unfortunately I can't find any tests for this file. We should stop the bleeding as fast as we can, so I'm filing a separate issue to cover this file with tests in a separate CL, especially given that it's not a trivial piece of code: []
	Change on 2015/02/24 by vanya <vanya@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=87099191