Always request the entire object for objects with Content-Encoding: gzip.
This CL applies the changes in [] back to Hadoop's GCS connector. Justification from that CL:

The size reported in the response header from GCS is the size of the compressed file.  GCSReadChannel uses the response header to set the size of the channel in openStreamAndSetSize() (which, in the case of gzip responses, will be less than the number of bytes that can be read) and the current position in the stream is updated in read() and other places.

Issue 1: read() fails after reading n > size bytes and retries.  currentPosition is used in the request header of openStreamAndSetSize.  This will cause an exception to be thrown in validatePosition and failure.
Issue 2: read() fails after reading n < size bytes and retries.  This will cause openStreamAndSetSize to request bytes n-... from GCS.  The returned suffix of the file will not be gzip encoded and the client library will throw an exception.

This CL fixes both of these by always requesting the full file from GCS and seeking to the correct position in the uncompressed stream that the [] client returns.
	Change on 2015/06/02 by mattlang <mattlang@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=95024908