Only re-try exceptions that are caused by SocketException or SocketTimeoutException, avoids re-try on SSLException directly.
Not all SSLExceptions are retryable. If we re-try the wrong type of exceptions, we will slow down the propagation of the root cause of the problem and waste resources.

Exceptions such as the one in [] is covered by looking into the cause.
	Change on 2015/09/03 by peihe <peihe@google.com>
-------------
Created by MOE: http://code.google.com/p/moe-java
MOE_MIGRATED_REVID=102285778