Add a new post-init helper script. (#74)
This initialization action can be used as either an initialization action or as a startup-script,
with the latter enabling support for commands which need to be re-run on reboot. This
initialization action adds polling for the cluster itself to be fully healthy before issuing
a command, effectively allowing "post-initiailization" actions which can include further setup
or submitting a job.