Close the readChannel when a lazy seek is initiated.
When performing a lazy seek before an in-place seek, the in-place seek will
perform a read on readChannel that may be at EOF. This prevents that by
pro-actively dumping the readChannel as soon as we initiate a lazy-seek.

As a lazy seek is pending, this will not require extra round trips as the next
read from the overall GoogleCloudStorageReadChannel will check if a lazy seek
is pending and if so, initiate the seek.

	Change on 2016/10/03 by angusdavis <angusdavis@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=134998995