When encountering errors in createEmptyObject, poll for objects for 3000ms.
When createing empty objects from map reduce and spark jobs, it's not uncommon
for there to be a race where multiple workers will attempt to create the same
empty object at the same time. It is also not uncommon for these workers to
receive a rate limit error when creating these objects. Further, it is not
uncommon for the empty object to not yet exist the first time we check if it
already exists.

This adds a polling loop into canIgnoreExceptionForEmptyObject to continually
check for object existance for up to 3 seconds.

The timeout for this polling loop can be configured via the setting:
"fs.gs.max.wait.for.empty.object.creation.ms" which is interpreted to be in
milliseconds. Polling can be disabled entirely by setting this value to be <= 0.

	Change on 2017/04/06 by angusdavis <angusdavis@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=152457820