Disable long-running flink yarn session by default.
The yarn session is currently configured to consume all YARN resources.
This makes it difficult to submit non-flink jobs to a cluster that has
been started with the flink init action.

This change disables the default session. Users can always fork the
script to change this setting or manually initiate their own sessions
out-of-band.