Use sed instead of cut to extract Spark settings (#151)
In /etc/spark/conf/spark-defaults.conf, some settings use "=" and other use
whitespace to define Spark's settings. With the Dataproc image
version upgrade, one of them changed to use whitespace and that
prevented this script from working properly. To add a bit of robustness
to the script, use regex to make sure this works both "=" and
whitespace. In addition, if spark.yarn.executor.memoryOverhead is not
defined, it uses the default Spark property value.

Fixes #131