Fix flink init action (#155)
The init action uses `set -e` to fail if any setup actions fail (i.e.,
exit non-zero). The script used to fetch metadata exits non-zero if the
attribute requested does not exist. We use a metadata attribute to
determine whether to start a detached yarn session. This causes the init
action to fail if the special attribute is not set (the default
behavior).