Allow users to specify a custom Flink snapshot URL
In order to support Beam use cases, customers need to be able to install
flink versions other than the ones supported by apt.  To facilitate
this, this commit adds the metadata key `flink-snapshot-url` which, when
set, will be used to download a flink tarball from a custom address.

This commit also cleans up the call pattern a little by separating
flink configuration and invocation.  This is useful if the yarn
application needs to be killed and restarted for any reason.