Add Beam Job Service init action (#344)
This sets up a Beam Job Service on the master node. It depends on Flink
and Docker being installed to function correctly.

The Beam team has advised us that Beam is not in a place where the portability
framework can be considered production ready, and as a result the stable
snapshots cannot be relied on. They advise instead that for the time being,
users build their own Job Service and Worker Image artifacts. This PR includes
both an init action for the job service, and an automated build script.
Build instructions are also available in the README.