Require openjdk to be authenticated
This commit removes the passing of the `--allow-unauthenticated` flag when installing the package `openjdk-8-jre-headless.

That flag was added in #620 to work around an issue in installing the hive and spark-python packages. Since openjdk-8-jre-headless was being installed in the same command, this flag was applied to it as well, even though it was not necessary.

This change locks that back down by splitting the openjdk-8-jre-headless package in a separate command that does not include the `--allow-unauthenticated` flag.