Initialize object before creating the channel in read channel tests
With FastFailOnNotFound (on by default), the object size is cached on
channel creation, which can cause errors if the size matters.

In actual use if the client is using strict generation consistency, then
the client will send the approriate generation in the request, and we
can rely on the server to never change the object size.

Change-Id: I76e3ec1c22400309ce2ddbc86744457ee0788f05