Add flag for Direct Path support
The enable_grpclb system property needs to be set for gRPC to use Direct
Path. The gRPC team has told me that this should be fine to set even if not
using Direct Path, so I think it makes the most sense to just set it in
the connector so that the users don't have to worry about it.

Change-Id: Iacbe8d5706056f81c0906cf9742994440781ce30