[spark-rapids.sh] Refactor NVIDIA driver installation for Rocky Linux 8 to use run file (#1359)
* Refactor NVIDIA driver installation for Rocky Linux 8

Updated the installation process for the NVIDIA GPU driver on Rocky Linux 8. The script now installs kernel development packages directly and downloads the CUDA installer run file, executing it in silent mode. The installer file is removed post-installation to clean up. This change simplifies the installation steps and ensures the correct driver version is used.

* feat: Enable spark-rapids on Dataproc 2.1 Rocky Linux 8

This commit integrates changes to enable the spark-rapids initialization action on Dataproc 2.1-rocky8 images.

- Updates the NVIDIA driver installation process in `spark-rapids.sh` for Rocky Linux:
  - Uses `curl` with retry and fail-fast options for downloading the CUDA installer.
  - Executes the NVIDIA installer with `--silent --driver --toolkit --no-opengl-libs` flags and wraps it in `execute_with_retries`.

- Modifies `test_spark_rapids.py` to enable tests for Rocky Linux on Dataproc 2.1 and below, while keeping them skipped for 2.2+ (Rocky 9).

This resolves the installation issues on Rocky 8. Further work is required to support Rocky 9 (Dataproc 2.2).

---------

Co-authored-by: C.J. Collier <cjac@google.com>